sy_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
channels==4.1.0
channels-redis
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent_kafka==2.9.0
cron-descriptor==1.4.3
diff-match-patch==20230430
django==5.0.6
django-admin-list-filters==1.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.2
django-import-export==4.0.9
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
humanize==4.10.0
kombu==5.3.7
openpyxl==3.1.4
paho-mqtt==2.1.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
psycopg2==2.9.9
pygame==2.6.0
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.2.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
supervisor==4.2.5
tablib==3.5.0
tornado==6.4.1
tzdata==2024.1
uvicorn[standard]
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
